// lib/db/rpc-registry.generated.ts
export type RpcArg = {
  name: string;
  type: string;
};

export type RpcSignature = {
  args: RpcArg[];
  returns: string;
  defaults: number;
};